+++ src/compas_viewer/scene/capsuleobject.py
@@ -17,4 +17,4 @@ class CapsuleObject(MeshObject):
     def __init__(self, capsule: Capsule, **kwargs):
         self.u = kwargs.get("u", int(2 * pi * capsule.radius / self.LINEARDEFLECTION))
 
-        super(CapsuleObject, self).__init__(mesh=Mesh.from_shape(capsule, u=self.u), **kwargs)
+        super().__init__(mesh=Mesh.from_shape(capsule, u=self.u), **kwargs)