+++ src/compas_viewer/scene/sphereobject.py
@@ -17,4 +17,4 @@ class SphereObject(MeshObject):
     def __init__(self, sphere: Sphere, **kwargs):
         self.u = kwargs.get("u", int(2 * pi * sphere.radius / self.LINEARDEFLECTION))
         self.v = kwargs.get("v", int(2 * pi * sphere.radius / self.LINEARDEFLECTION))
-        super(SphereObject, self).__init__(mesh=Mesh.from_shape(sphere, u=self.u, v=self.v), **kwargs)
+        super().__init__(mesh=Mesh.from_shape(sphere, u=self.u, v=self.v), **kwargs)